Make the FingerprintDialog sample buildable with Android.mk so that both engineers can quickly iterate the changes over the sample.
From the top level of the Android source directory, now the sample can be built
by '$ make FingerprintDialog'.

This CL deletes the dagger dependency and the unit test depends on dagger,
but dagger wasn't relevant to the usage of the FingerprintManager and the
unit test wasn't pushed to GitHub. So should be no impact.

Also the CL makes the sample not depend on the Application/template directory
because otherwise './gradlew preflight' command is needed to build the sample
by 'make FingerprintDialog', which isn't good experience.

Bug: 27369902
Change-Id: If37836c1a51cbf1b8bede81210092edec41b6926