Some polish for DragAndDropAcrossApp sample
1. Do not re-create an image file if it already exists.
(this prevents a nasty race when DropTarget will try to read
a partially written file when restoring the image on
an activity resize).

2. Do not use ClipDescription.MIMETYPE_TEXT_URILIST
with a "content" scheme URI, use the actual MIME type
instead (this is what JavaDoc explicitly requires).

3. Do not call requestDragAndDrop permissions if
a dropped content is not a "content" scheme URI.

4. Add a checkbox to control permissions release explicitly.
Without it the image restore behavior is confusing.

5. Remove mention of "text" from the explanatory message.

Change-Id: Iedd1306687286d051a3a1b768f95c800fb3d5cd4