Initial commit for ContentProviderPaging.
Starting from Android O, ContentProvider and DocumentsProvider now
supports paging instead of fetching all the data at once.
This sample demonstrates how to use it by fetching the part of images from the
local storage.
As the user scrolls to the last fetched item, the next images are going
to be fetched.

The apk is uploaded at
https://drive.google.com/open?id=0B0A2UDVH_Z4CdUJ3QkludHJNZTA

Test: Manual

Change-Id: I391d6808ce6ae0093e9ccf08995d7a442cdf5fdd