Use PhoneDeviceType API in WearVerifyRemoteApp.
This allows the example app to work correctly with Android Wear apps
globally, previously it did not work for Android Wear for China.

The new getPhoneDeviceType API added in Wearable Support 2.1.0 allows
for determining if the paired phone is Android or iOS which is a more
accurate way to determine if a market or iTunes App Store URI is
appropriate for the remote intent.

Bug: 35661437
Change-Id: I70888391d59c060f8e40e2c929f5e2c64b07f647