Add maintainer's note to AutofillFramework
The Kotlin version of this sample is temporarily out of date due to
API changes. Add a maintainer's note to let users know that the Java
version is more current at the moment.

It looks like this change was made to Github directly, but will be
wiped out by the next push.

Also, it was a <h1> element, which means the samples parser is
interpreting it as the title of the sample.

Test: N/A
Bug: 73495046
Change-Id: I41e56bcf511b6a238bc32ebbc628d581a8e4e584